# api/openapi.yaml
openapi: 3.0.3
info:
  title: ASM Platform API
  version: 0.8.0
  description: |
    Contract for the HTTP routes currently mounted by cmd/api. Discovery scope,
    task template, and task run management are implemented below service level
    but are not listed here until HTTP handlers are added.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /projects/{project_id}/tickets:
    get:
      tags: [tickets]
      summary: List tickets.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/PageNumber"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Paged tickets.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
    post:
      tags: [tickets]
      summary: Create a ticket.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "201":
          description: Created ticket.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
  /projects/{project_id}/tickets/{id}:
    get:
      tags: [tickets]
      summary: Get ticket detail.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Ticket detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
  /projects/{project_id}/tickets/{id}/status-transitions:
    post:
      tags: [tickets]
      summary: Transition a ticket status.
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Updated ticket.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectID:
      name: project_id
      in: path
      required: true
      schema: { type: integer, format: uint64, minimum: 1 }
...
    ID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: uint64, minimum: 1 }
...
    PageNumber:
      name: page_number
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  schemas:
    GenericEnvelope:
      type: object
      required: [request_id, data]
      properties:
        request_id: { type: string }
        data: {}
...